Splits animation into four frames called separately.
This allows the descriptors to change mid-spin, without
interrupting the animation. In addition, it prevents the
setTimeout that keeps track of the descriptorChange and the
animation css (infinite) to fall out of sync, which happened after
a few minutes (since JS is single-threaded and can only perform one
action at a time).