Solves glitch in Prism animation.
A slight delay between when one animation finished and the next started
caused the Prism to revert back to its original position with face1
visible. By setting animation-fill-mode to 'forwards' each animation
stays in its place after finishing, so this 'glitch-like' behavior is
avoided.

In addition, in the Schema, view links can have newTab set to 'true'
which causes the link target to be '_blank', thus opening the view in
a new tab.